Random: add Fixed of string funcs
