+++ random/const.go
@@ -1,1 +1,5 @@
 package random
+
+const (
+	alphaNumerics = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
+)
